refactor: extract section renderers and shared utilities
- Extract renderColors and renderTypography into dedicated section modules
- Move deepMerge to shared util, remove duplicate implementation from CssLoader
- Add PresetOptions type and createAddonOptions helper to preset.ts to
  eliminate duplicated AddonOptions construction
