+++ src/addon/core/theme-transformer/CsfGenerator.ts
@@ -2,6 +2,7 @@ import { Color } from './Color';
 import { Typography } from '../../types';
 import { AddonOptions } from './AddonOptions';
 import { sanitizeExportName } from '../../util';
+import { renderColors, renderTypography } from './sections';
 
 export class CsfGenerator {
     private addonOptions: AddonOptions;
@@ -139,7 +140,7 @@ export class CsfGenerator {
         render: () => (
             <Wrapper>
                 <Container>
-                    ${this.renderColors(colors)}
+                    ${renderColors(colors)}
                 </Container>
             </Wrapper>
         )
@@ -154,7 +155,7 @@ export class CsfGenerator {
         render: () => (
             <Wrapper>
                 <Container>
-                    ${this.renderTypography(typography)}
+                    ${renderTypography(typography)}
                 </Container>
             </Wrapper>
         )
@@ -174,10 +175,10 @@ export class CsfGenerator {
         const elements: string[] = [];
         enabledSections.forEach((section, idx) => {
             if (section === 'Colors') {
-                elements.push(this.renderColors(colors));
+                elements.push(renderColors(colors));
             }
             if (section === 'Typography') {
-                elements.push(this.renderTypography(typography));
+                elements.push(renderTypography(typography));
             }
             if (idx < enabledSections.length - 1) {
                 elements.push('<HorizontalRule />');
@@ -197,80 +198,4 @@ export class CsfGenerator {
         `;
     }
 
-    private renderColors(colors: Color[]): string {
-        const hasColors = colors.length > 0;
-        return `
-        <Title>Colors</Title>
-        <br />
-        ${
-            hasColors
-                ? `
-        <ColorPalette>
-            {${JSON.stringify(colors)}.map((color) => (
-                <ColorItem
-                    key={color.baseName}
-                    title={color.baseName}
-                    subtitle={color.subtitle}
-                    colors={color.shades}
-                />
-            ))}
-        </ColorPalette>
-        `
-                : `
-        <NoneDetectedText>
-            No colors detected. To see a color, add it to your Tailwind configuration, or ensure Tailwind's defaults are not being overridden.
-        </NoneDetectedText>
-        `
-        }
-        `;
-    }
-
-    private renderTypography(typography: Typography): string {
-        const fontSizes = Object.values(typography.size);
-        const fontWeights = Object.entries(typography.weight);
-        const fontFamilies = Object.entries(typography.type);
-        const sampleText =
-            'Lorem ipsum dolor sit amet, consectetur adipiscing elit.';
-        const hasFontFamily = fontFamilies.length > 0; // TODO: Update this handling so that it handles just weight and/or just size
-
-        return `
-        <Title>Typography</Title>
-        <br />
-        ${
-            hasFontFamily
-                ? `
-        {${JSON.stringify(fontFamilies)}.map(([label, fontFamily], familyIndex) => (
-            <div key={label}>
-                <FontHeaderSection>
-                    <div>
-                        <b>Font Face: </b>
-                        <span style={{ fontFamily }}>{label}</span>
-                    </div>
-                    <div>
-                        <b>Weights: </b>
-                        {${JSON.stringify(fontWeights)}.map(([weightLabel, weightValue], index) => (
-                            <span key={weightLabel} style={{ fontWeight: weightValue, fontFamily }}>
-                                {\`\${weightValue}(\${weightLabel})\${index < ${fontWeights.length} - 1 ? ', ' : ''}\`}
-                            </span>
-                        ))}
-                    </div>
-                </FontHeaderSection>
-                <Typeset
-                    fontSizes={${JSON.stringify(fontSizes)}}
-                    fontWeight={400}
-                    sampleText="${sampleText}"
-                    fontFamily={fontFamily}
-                />
-                {familyIndex < ${fontFamilies.length} - 1 && <HorizontalRule />}
-            </div>
-        ))}
-        `
-                : `
-        <NoneDetectedText>
-            No font families detected. To see typography, add a font family to your Tailwind configuration, or ensure Tailwind's defaults are not being overridden.
-        </NoneDetectedText>
-        `
-        }
-        `;
-    }
 }
