+++ src/addon/core/theme-transformer/sections/index.ts
@@ -1,0 +1,2 @@
+export { renderColors } from './renderColors';
+export { renderTypography } from './renderTypography';
