+++ src/addon/util/index.ts
@@ -1,3 +1,4 @@
 export { Logger } from './Logger';
 export { UnitConverter } from './UnitConverter';
 export { sanitizeExportName } from './sanitizeExportName';
+export { deepMerge } from './deepMerge';
